+++ gamelib/include/bq/graphics/animation.h
@@ -1,16 +1,17 @@
 #pragma once
-#include <bq/common.h>
-#include <bq/resource/resource_holder.h>
+
+
 namespace bq {
+	
 	class animation {
 
 		sf::Sprite m_sprite;
-		int startingy, spritesize, numsprites;
-		int index = 0;
-		int frames = 0;
-		int frames_to_update = 0;
+		size_t startingy, spritesize, numsprites;
+		size_t index = 0;
+		size_t frames = 0;
+		size_t frames_to_update = 0;
 	public:
-		animation(std::string, int, int, int, int);
+		animation(std::string, size_t, size_t, size_t, size_t);
 		void update();
 		sf::Sprite& get();
 	};
