+++ gamelib/include/bq/util/block_collision_effects.h
@@ -3,7 +3,7 @@
 namespace bq {
 	struct block_collision_effects {
 		bool m_collision = false;
-		int m_damage = 0;
+		size_t m_damage = 0;
 		float m_slow = 0.f;
 	};
 }
